shopping_cart/requirements.txt
# Django and REST Framework
Django>=4.2.0,<5.0.0
djangorestframework>=3.14.0

# Database
psycopg2-binary>=2.9.9

# Testing
pytest>=8.0.0
pytest-django>=4.7.0

# Code Quality
flake8>=6.0.0

# WSGI Server
gunicorn>=21.2.0

# Static file serving
whitenoise>=6.6.0
